+++ src/Http/Middleware/CheckCredentials.php
@@ -3,7 +3,7 @@
 namespace Laravel\Passport\Http\Middleware;
 
 use Closure;
-use Illuminate\Auth\AuthenticationException;
+use Laravel\Passport\Exceptions\AuthenticationException;
 use Laravel\Passport\TokenRepository;
 use League\OAuth2\Server\Exception\OAuthServerException;
 use League\OAuth2\Server\ResourceServer;
@@ -40,6 +40,21 @@ abstract class CheckCredentials
     }
 
     /**
+     * Specify the scopes for the middleware.
+     *
+     * @param  array|string  $scopes
+     * @return string
+     */
+    public static function using(...$scopes)
+    {
+        if (is_array($scopes[0])) {
+            return static::class.':'.implode(',', $scopes[0]);
+        }
+
+        return static::class.':'.implode(',', $scopes);
+    }
+
+    /**
      * Handle an incoming request.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -47,7 +62,7 @@ abstract class CheckCredentials
      * @param  mixed  ...$scopes
      * @return mixed
      *
-     * @throws \Illuminate\Auth\AuthenticationException
+     * @throws \Laravel\Passport\Exceptions\AuthenticationException
      */
     public function handle($request, Closure $next, ...$scopes)
     {
@@ -93,7 +108,7 @@ abstract class CheckCredentials
      * @param  \Laravel\Passport\Token  $token
      * @return void
      *
-     * @throws \Illuminate\Auth\AuthenticationException
+     * @throws \Laravel\Passport\Exceptions\AuthenticationException
      */
     abstract protected function validateCredentials($token);