+++ src/Http/Middleware/CheckForAnyScope.php
@@ -2,11 +2,26 @@
 
 namespace Laravel\Passport\Http\Middleware;
 
-use Illuminate\Auth\AuthenticationException;
+use Laravel\Passport\Exceptions\AuthenticationException;
 use Laravel\Passport\Exceptions\MissingScopeException;
 
 class CheckForAnyScope
 {
+    /**
+     * Specify the scopes for the middleware.
+     *
+     * @param  array|string  $scopes
+     * @return string
+     */
+    public static function using(...$scopes)
+    {
+        if (is_array($scopes[0])) {
+            return static::class.':'.implode(',', $scopes[0]);
+        }
+
+        return static::class.':'.implode(',', $scopes);
+    }
+
     /**
      * Handle the incoming request.
      *
@@ -15,7 +30,7 @@ class CheckForAnyScope
      * @param  mixed  ...$scopes
      * @return \Illuminate\Http\Response
      *
-     * @throws \Illuminate\Auth\AuthenticationException|\Laravel\Passport\Exceptions\MissingScopeException
+     * @throws \Laravel\Passport\Exceptions\AuthenticationException|\Laravel\Passport\Exceptions\MissingScopeException
      */
     public function handle($request, $next, ...$scopes)
     {