+++ src/PassportServiceProvider.php
@@ -16,8 +16,9 @@ use Laravel\Passport\Bridge\PersonalAccessGrant;
 use Laravel\Passport\Bridge\RefreshTokenRepository;
 use Laravel\Passport\Guards\TokenGuard;
 use Laravel\Passport\Http\Controllers\AuthorizationController;
-use Lcobucci\JWT\Configuration;
-use Lcobucci\JWT\Parser;
+use Lcobucci\JWT\Encoding\JoseEncoder;
+use Lcobucci\JWT\Parser as ParserContract;
+use Lcobucci\JWT\Token\Parser;
 use League\OAuth2\Server\AuthorizationServer;
 use League\OAuth2\Server\CryptKey;
 use League\OAuth2\Server\Grant\AuthCodeGrant;
@@ -294,8 +295,8 @@ class PassportServiceProvider extends ServiceProvider
      */
     protected function registerJWTParser()
     {
-        $this->app->singleton(Parser::class, function () {
-            return Configuration::forUnsecuredSigner()->parser();
+        $this->app->singleton(ParserContract::class, function () {
+            return new Parser(new JoseEncoder);
         });
     }