+++ tests/Feature/ActingAsTest.php
@@ -3,12 +3,11 @@
 namespace Laravel\Passport\Tests\Feature;
 
 use Illuminate\Contracts\Routing\Registrar;
-use Illuminate\Foundation\Auth\User;
 use Illuminate\Support\Facades\Route;
-use Laravel\Passport\HasApiTokens;
 use Laravel\Passport\Http\Middleware\CheckForAnyScope;
 use Laravel\Passport\Http\Middleware\CheckScopes;
 use Laravel\Passport\Passport;
+use Workbench\App\Models\User;
 
 class ActingAsTest extends PassportTestCase
 {
@@ -23,7 +22,7 @@ class ActingAsTest extends PassportTestCase
             return 'bar';
         })->middleware('auth:api');
 
-        Passport::actingAs(new PassportUser());
+        Passport::actingAs(new User());
 
         $response = $this->get('/foo');
         $response->assertSuccessful();
@@ -41,11 +40,26 @@ class ActingAsTest extends PassportTestCase
             return 'bar';
         })->middleware(CheckScopes::class.':admin,footest');
 
-        Passport::actingAs(new PassportUser(), ['admin', 'footest']);
+        Passport::actingAs(new User(), ['admin', 'footest']);
 
         $response = $this->get('/foo');
         $response->assertSuccessful();
         $response->assertSee('bar');
+    }
+
+    public function testItCanGenerateDefinitionViaStaticMethod()
+    {
+        $signature = (string) CheckScopes::using('admin');
+        $this->assertSame('Laravel\Passport\Http\Middleware\CheckScopes:admin', $signature);
+
+        $signature = (string) CheckScopes::using('admin', 'footest');
+        $this->assertSame('Laravel\Passport\Http\Middleware\CheckScopes:admin,footest', $signature);
+
+        $signature = (string) CheckForAnyScope::using('admin');
+        $this->assertSame('Laravel\Passport\Http\Middleware\CheckForAnyScope:admin', $signature);
+
+        $signature = (string) CheckForAnyScope::using('admin', 'footest');
+        $this->assertSame('Laravel\Passport\Http\Middleware\CheckForAnyScope:admin,footest', $signature);
     }
 
     public function testActingAsWhenTheRouteIsProtectedByCheckForAnyScopeMiddleware()
@@ -59,7 +73,7 @@ class ActingAsTest extends PassportTestCase
             return 'bar';
         })->middleware(CheckForAnyScope::class.':admin,footest');
 
-        Passport::actingAs(new PassportUser(), ['footest']);
+        Passport::actingAs(new User(), ['footest']);
 
         $response = $this->get('/foo');
         $response->assertSuccessful();
@@ -76,7 +90,7 @@ class ActingAsTest extends PassportTestCase
             return 'bar';
         });
 
-        Passport::actingAs(new PassportUser(), ['foo', 'baz']);
+        Passport::actingAs(new User(), ['foo', 'baz']);
 
         $response = $this->get('/foo');
         $response->assertSuccessful();
@@ -93,17 +107,10 @@ class ActingAsTest extends PassportTestCase
             return 'bar';
         });
 
-        Passport::actingAs(new PassportUser(), ['foo']);
+        Passport::actingAs(new User(), ['foo']);
 
         $response = $this->get('/foo');
         $response->assertSuccessful();
         $response->assertSee('bar');
     }
-}
-
-class PassportUser extends User
-{
-    use HasApiTokens;
-
-    protected $table = 'users';
 }