+++ tests/Feature/PassportTestCase.php
@@ -3,14 +3,15 @@
 namespace Laravel\Passport\Tests\Feature;
 
 use Illuminate\Contracts\Config\Repository;
-use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Foundation\Testing\LazilyRefreshDatabase;
 use Laravel\Passport\Passport;
-use Laravel\Passport\PassportServiceProvider;
+use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase;
+use Workbench\App\Models\User;
 
 abstract class PassportTestCase extends TestCase
 {
-    use RefreshDatabase;
+    use LazilyRefreshDatabase, WithWorkbench;
 
     const KEYS = __DIR__.'/../keys';
     const PUBLIC_KEY = self::KEYS.'/oauth-public.key';
@@ -18,50 +19,32 @@ abstract class PassportTestCase extends TestCase
 
     protected function setUp(): void
     {
+        $this->afterApplicationCreated(function () {
+            Passport::loadKeysFrom(self::KEYS);
+
+            @unlink(self::PUBLIC_KEY);
+            @unlink(self::PRIVATE_KEY);
+
+            $this->artisan('passport:keys');
+        });
+
+        $this->beforeApplicationDestroyed(function () {
+            @unlink(self::PUBLIC_KEY);
+            @unlink(self::PRIVATE_KEY);
+        });
+
         parent::setUp();
-
-        $this->artisan('migrate:fresh');
-
-        Passport::loadKeysFrom(self::KEYS);
-
-        @unlink(self::PUBLIC_KEY);
-        @unlink(self::PRIVATE_KEY);
-
-        $this->artisan('passport:keys');
     }
 
-    protected function getEnvironmentSetUp($app)
+    protected function defineEnvironment($app)
     {
         $config = $app->make(Repository::class);
 
-        $config->set('auth.defaults.provider', 'users');
-
-        if (($userClass = $this->getUserClass()) !== null) {
-            $config->set('auth.providers.users.model', $userClass);
-        }
-
-        $config->set('auth.guards.api', ['driver' => 'passport', 'provider' => 'users']);
-
-        $app['config']->set('database.default', 'testbench');
-
-        $app['config']->set('database.connections.testbench', [
-            'driver'   => 'sqlite',
-            'database' => ':memory:',
-            'prefix'   => '',
+        $config->set([
+            'auth.defaults.provider' => 'users',
+            'auth.providers.users.model' => User::class,
+            'auth.guards.api' => ['driver' => 'passport', 'provider' => 'users'],
+            'database.default' => 'testing',
         ]);
-    }
-
-    protected function getPackageProviders($app)
-    {
-        return [PassportServiceProvider::class];
-    }
-
-    /**
-     * Get the Eloquent user model class name.
-     *
-     * @return string|null
-     */
-    protected function getUserClass()
-    {
     }
 }